Fix content issues. See full description.
In case of ActiveRecord related object, when nodes content_id is
missing, or when related object was shomehow deleted from DB, when user
opens node it is rendered, and can be saved, however there are no fields
visible for related model.

From users perspective you can't do anything to fix this. Site might be
broken because of this.

This "fix" ensures that fields for related AR object is always rendered.
When user saves node, object will be created, thus any site specific
issue can be fixed by user.

It is not possible to fix this kind of issue with database migration
(i.e. finding nodes with are related to AR objects, that can't be
found), because of custom validations that related object can have.

The only way to fix DB is such state (missing related object) is for
user to update node manually and fill in all required info.