Make it possible to open hyperlinks from MS office
links need to point to new sessions path with redirect_to param, which
must point to target path, where to redirect

The problem is that MS office tries to download target page, but it's
not authorized, so it's redirected to login page. Then MS office opens
browser, but instead of passing original URL it passes redirect URL, so
even if you're logged in, you can't open link from office poiting to
correct resource.