Fix case when profile controller is inherited and used different resource_clas
When controller then extends profile controller sets resource class to
use sublcass of Releaf::Permissions::User, it's actually not being used,
because @resource is set in #setup to current user.

To fix this, we specifically say which class we want to work with by
using #becomes.

After this change if you add some custom validation to subclass of
Releaf::Permissions::User and override .resource_class to return it.
It will actually run these additional validations.