Invert the method of detecting whether a toolbox has been initialized
Instead of initially rendering the toolbox button as disabled in HTML
until the script initalizes it, render the trigger as enabled but
only consider the toolbox to be fully initialized once the 'initialized'
class hass been added by the script.

This is already how the dialogs get marked as having finished opening.