Cache content class names as strings instead of class constants
Caching node configuration classes as constants caused problems
in development environment when application classes were being reloaded.

The releaf classes are not reloaded and if they internally store
a reference to an application class that gets automatically reloaded,
the reference becomes outdated when application code is being edited.

Storing the class names as strings and resolving to constants on demand
seems to solve this problem.