Fix has_error? matcher
There was a bug that caused the matcher to sometimes return false positives when
used with the :field option. It would incorrectly return true if the
expected field was found and the expected error message was present
anywhere on the form, including cases when the error was on a different field.